nfs: fix compilation warning
Patchwork-id: 35722
O-Subject: [RHEL6.2 PATCH 06/23] nfs: fix compilation warning
Bugzilla: 701373
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Jovi Zhang <bookjovi@gmail.com>

this commit fix compilation warning as following:
linux-2.6/fs/nfs/nfs4proc.c:3265: warning: comparison of distinct pointer types lacks a cast

+++ fs/nfs/nfs4proc.c
@@ -3256,7 +3256,7 @@ static int buf_to_pages_noslab(const void *buf, size_t buflen,
 	spages = pages;
 
 	do {
-		len = min_t(int, PAGE_CACHE_SIZE, buflen);
+		len = min_t(size_t, PAGE_CACHE_SIZE, buflen);
 		newpage = alloc_page(GFP_KERNEL);
 
 		if (newpage == NULL)
